refactoring regulation list
