+++ app.js
@@ -26,8 +26,9 @@ var newPersonEntity;
 var consolidatedList = [];
 
 var ccas_list = require("./api/ccas").masterListJSON;
+var reg_list = require("./api/regList").regulationListLink;
 
-consolidatedList = [...ccas_list.jsonList];
+consolidatedList = [...ccas_list.jsonList,...reg_list.jSONList];
 
 let lteListData = [];
 
@@ -75,52 +76,52 @@ axios({
 	consolidatedList = [...lteListData];
 });
 
-let regulationFilteredData = [];
+// let regulationFilteredData = [];
 
-const regulationListLink = "https://laws-lois.justice.gc.ca/eng/XML/SOR-2002-284.xml";
+// const regulationListLink = "https://laws-lois.justice.gc.ca/eng/XML/SOR-2002-284.xml";
 
-//GET Regulation List
-axios({
-	method: "get",
-	url: regulationListLink,
-}).then(function (response) {
-	let fetchData = JSON.parse(xmlParser.toJson(response.data));
-	//console.log(response);
-	let filteredList = fetchData["Regulation"]["Body"]["Section"][0]["List"]["Item"];
+// //GET Regulation List
+// axios({
+// 	method: "get",
+// 	url: regulationListLink,
+// }).then(function (response) {
+// 	let fetchData = JSON.parse(xmlParser.toJson(response.data));
+// 	//console.log(response);
+// 	let filteredList = fetchData["Regulation"]["Body"]["Section"][0]["List"]["Item"];
 
-	for (let i = 0; i < filteredList.length; i++) {
-		let temp = filteredList[i]["Text"];
-		let text = temp.toString();
-		let group_name = text.split(" (also")[0];
-		let description = text.split(" (also")[1];
-		regulationFilteredData[i] = {
-			name: group_name,
-			date: filteredList[0]["lims:inforce-start-date"],
-			link: regulationListLink,
-			address: {
-				streetNum: faker.datatype.number(),
-				street: faker.address.streetName(),
-				city: faker.address.cityName(),
-				prov: faker.address.state(),
-				postal: faker.address.zipCodeByState(),
-			},
-			id: {
-				idType: "Driver's License",
-				idNumber: faker.finance.routingNumber(),
-				idJuristiction: faker.address.state(),
-			},
-			accountInfo: {
-				locale: "Domestic",
-				institution: "TD",
-				transitNum: faker.finance.routingNumber(),
-				accountNum: faker.finance.account(),
-				status: "Active",
-			},
-			dateRange: "2020-2021"
-		};
-	}
-	consolidatedList = [...regulationFilteredData];
-});
+// 	for (let i = 0; i < filteredList.length; i++) {
+// 		let temp = filteredList[i]["Text"];
+// 		let text = temp.toString();
+// 		let group_name = text.split(" (also")[0];
+// 		let description = text.split(" (also")[1];
+// 		regulationFilteredData[i] = {
+// 			name: group_name,
+// 			date: filteredList[0]["lims:inforce-start-date"],
+// 			link: regulationListLink,
+// 			address: {
+// 				streetNum: faker.datatype.number(),
+// 				street: faker.address.streetName(),
+// 				city: faker.address.cityName(),
+// 				prov: faker.address.state(),
+// 				postal: faker.address.zipCodeByState(),
+// 			},
+// 			id: {
+// 				idType: "Driver's License",
+// 				idNumber: faker.finance.routingNumber(),
+// 				idJuristiction: faker.address.state(),
+// 			},
+// 			accountInfo: {
+// 				locale: "Domestic",
+// 				institution: "TD",
+// 				transitNum: faker.finance.routingNumber(),
+// 				accountNum: faker.finance.account(),
+// 				status: "Active",
+// 			},
+// 			dateRange: "2020-2021"
+// 		};
+// 	}
+// 	consolidatedList = [...regulationFilteredData];
+// });
 
 var finalList = [];
